code seperation done
