+++ tests/test_historical_data.py
@@ -27,10 +27,10 @@ def alpaca():
 
 
 ########################################################
-# Test cases for PyAlpacaApi.get_stock_historical_data #
+# Test cases for PyAlpacaApi.get_stock_data #
 ########################################################
-def test_get_stock_historical_data_1d(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1d")
+def test_get_stock_data_1d(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1d")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
@@ -44,8 +44,8 @@ def test_get_stock_historical_data_1d(alpaca):
     assert isinstance(stock_data["symbol"][0], str)
 
 
-def test_get_stock_historical_data_1w(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1w")
+def test_get_stock_data_1w(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1w")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
@@ -59,8 +59,8 @@ def test_get_stock_historical_data_1w(alpaca):
     assert isinstance(stock_data["symbol"][0], str)
 
 
-def test_get_stock_historical_data_1m(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1m")
+def test_get_stock_data_1m(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1m")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
@@ -74,8 +74,8 @@ def test_get_stock_historical_data_1m(alpaca):
     assert isinstance(stock_data["symbol"][0], str)
 
 
-def test_get_stock_historical_data_5m(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="5m")
+def test_get_stock_data_5m(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="5m")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
@@ -89,8 +89,8 @@ def test_get_stock_historical_data_5m(alpaca):
     assert isinstance(stock_data["symbol"][0], str)
 
 
-def test_get_stock_historical_data_15m(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="15m")
+def test_get_stock_data_15m(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="15m")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
@@ -104,8 +104,8 @@ def test_get_stock_historical_data_15m(alpaca):
     assert isinstance(stock_data["symbol"][0], str)
 
 
-def test_get_stock_historical_data_30m(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="30m")
+def test_get_stock_data_30m(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="30m")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
@@ -119,8 +119,8 @@ def test_get_stock_historical_data_30m(alpaca):
     assert isinstance(stock_data["symbol"][0], str)
 
 
-def test_get_stock_historical_data_1h(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1h")
+def test_get_stock_data_1h(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="1h")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
@@ -134,8 +134,8 @@ def test_get_stock_historical_data_1h(alpaca):
     assert isinstance(stock_data["symbol"][0], str)
 
 
-def test_get_stock_historical_data_4h(alpaca):
-    stock_data = alpaca.get_stock_historical_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="4h")
+def test_get_stock_data_4h(alpaca):
+    stock_data = alpaca.history.get_stock_data(symbol="AAPL", start=month_ago, end=previous_day, timeframe="4h")
     assert not stock_data.empty
     assert stock_data["symbol"][0] == "AAPL"
     assert isinstance(stock_data["close"][0], float)
